+++ src/main/kotlin/com/emberjs/cli/EmberCliFrameworkDetector.kt
@@ -1,5 +1,6 @@
 package com.emberjs.cli
 
+import com.emberjs.EmberTagNameProvider
 import com.emberjs.utils.NotLibrary
 import com.intellij.framework.FrameworkType
 import com.intellij.framework.detection.DetectedFrameworkDescription
@@ -7,6 +8,8 @@ import com.intellij.framework.detection.FileContentPattern
 import com.intellij.framework.detection.FrameworkDetectionContext
 import com.intellij.framework.detection.FrameworkDetector
 import com.intellij.ide.projectView.actions.MarkRootActionBase
+import com.intellij.lang.javascript.JavaScriptFileType
+import com.intellij.lang.javascript.frameworks.react.ReactXmlExtension
 import com.intellij.openapi.fileTypes.FileType
 import com.intellij.openapi.fileTypes.PlainTextFileType
 import com.intellij.openapi.module.ModuleUtilCore
@@ -17,20 +20,25 @@ import com.intellij.openapi.roots.ui.configuration.ModulesProvider
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.patterns.ElementPattern
 import com.intellij.util.indexing.FileContent
+import com.intellij.xml.DefaultXmlExtension
+import com.intellij.xml.XmlExtension
+import com.intellij.xml.XmlTagNameProvider
 
 class EmberCliFrameworkDetector : FrameworkDetector("Ember") {
     /** The `.ember-cli` file is detected as plain text */
-    override fun getFileType(): FileType = PlainTextFileType.INSTANCE
+    override fun getFileType(): FileType = JavaScriptFileType.INSTANCE
 
     override fun createSuitableFilePattern(): ElementPattern<FileContent> {
         return FileContentPattern.fileContent()
-                .withName(".ember-cli")
+                .withName("ember-cli-build.js")
                 .with(NotLibrary)
     }
 
     override fun getFrameworkType(): FrameworkType = EmberFrameworkType
 
     override fun detect(newFiles: MutableCollection<out VirtualFile>, context: FrameworkDetectionContext): MutableList<out DetectedFrameworkDescription> {
+        XmlExtension.EP_NAME.point.unregisterExtensions({it !is DefaultXmlExtension })
+        XmlTagNameProvider.EP_NAME.point.unregisterExtensions({ it !is ReactXmlExtension && it !is EmberTagNameProvider })
         val rootDir = newFiles.firstOrNull()?.parent
         if (rootDir != null && !isConfigured(newFiles, context.project)) {
             return mutableListOf(EmberFrameworkDescription(rootDir, newFiles))