+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Changelog
 ===============================================================================
+## 2022.1.8
+- fix: null access violation
+- fix: detect framework with ember-cli-build.js
+
 ## 2022.1.7
 - fix: internal components did not have a reference